% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_UVMR.R
\name{stand_UVMR}
\alias{stand_UVMR}
\title{标准单变量孟德尔随机化分析}
\usage{
stand_UVMR(
  expgwas,
  outgwas,
  clump_p1 = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  pop = "EUR",
  outfile = "MR结果",
  steiger = T,
  Fvalue = T,
  pt = T
)
}
\arguments{
\item{expgwas}{输入暴露的GWAS摘要数据}

\item{outgwas}{输入结局的GWAS摘要数据}

\item{clump_p1}{输入工具变量的选择P值,默认5e-08}

\item{clump_r2}{输入工具变量的选择的r2,默认0.001}

\item{clump_kb}{输入工具变量的选择的距离,默认10000}

\item{pop}{输入工具变量的选择的人群,默认EUR}

\item{outfile}{输入分析结果的文件夹}

\item{steiger}{是否进行反向过滤，默认是TURE}

\item{Fvalue}{是否计算F值，默认是TURE}

\item{pt}{是否进行绘图，默认是TURE}
}
\description{
标准单变量孟德尔随机化分析
}
